POCKET/requirements.txt
Django
djangorestframework
mysqlclient
gunicorn
Pillow
django-simple-history


